edit car controll components
