+++ src/app/products/arm/[deviceId]/armJoyStick.tsx
@@ -8,9 +8,10 @@ type Props = {
     client : MqttClient | null;
     isConnected:boolean;
     topic:string;
+    onLogReturn: (data: string) => void;
 };
 
-export default function ArmJoyStick({device_id , isLoading ,client,isConnected,topic}:Props) {
+export default function ArmJoyStick({device_id , isLoading ,client,isConnected,topic,onLogReturn,}:Props) {
     const [fw, setFW] = useState<boolean>(false);
     const [bw, setBW] = useState<boolean>(false);
     const [lf, setLF] = useState<boolean>(false);
@@ -23,36 +24,43 @@ export default function ArmJoyStick({device_id , isLoading ,client,isConnected,t
         if (client && isConnected) {
             client.publish(topic, "ctrl/forward");
             console.log("forward : ON");
+            onLogReturn("forward : ON")
         }
       };
       const Backward = async () => {
         if (client && isConnected) {
             client.publish(topic, "ctrl/backward");
             console.log("backward : ON");
+            onLogReturn("backward : ON")
+
         }
       };
       const Left = async () => {
         if (client && isConnected) {
             client.publish(topic, "ctrl/left");
             console.log("left : ON");
+            onLogReturn("left : ON")
         }
       };
       const Right = async () => {
         if (client && isConnected) {
             client.publish(topic, "ctrl/right");
             console.log("right : ON");
+            onLogReturn("right : ON")
         }
       };
       const Up = async () => {
         if (client && isConnected) {
             client.publish(topic, "ctrl/up");
             console.log("up : ON");
+            onLogReturn("up : ON")
         }
       };
       const Down = async () => {
         if (client && isConnected) {
             client.publish(topic, "ctrl/down");
             console.log("down : ON");
+            onLogReturn("down : ON");
         }
       };
       const Hold = () => {
@@ -61,10 +69,12 @@ export default function ArmJoyStick({device_id , isLoading ,client,isConnected,t
             client.publish(topic, "ctrl/holdON");
             setHold(true);
             console.log("hold : ON");
+            onLogReturn("hold : ON")
           } else {
             client.publish(topic, "ctrl/holdOFF");
             setHold(false);
             console.log("hold : OFF");
+            onLogReturn("hold : OFF")
           }
         }
       };
@@ -82,11 +92,11 @@ export default function ArmJoyStick({device_id , isLoading ,client,isConnected,t
                 </div>
                 <div className={`duration-1000 grid place-items-center shadow-md shadow-gray-900  rounded-full ${isLoading? 'bg-blue-800':'bg-gray-500'} `}>
                     {hold == false ? (
-                        <button className={`text-4xl duration-75 bg-gray-800 active:bg-violet-700  shadow-md shadow-gray-900 text-white  rounded-full py-8 px-2 ${isLoading? 'animate-fadeIn':'opacity-0'} `} onClick={Hold}>
+                        <button className={`active:shadow-inner active:shadow-black  text-4xl duration-75 bg-gray-800 active:bg-violet-700  shadow-md shadow-gray-900 text-white  rounded-full py-8 px-2 ${isLoading? 'animate-fadeIn':'opacity-0'} `} onClick={Hold}>
                         HOLD
                     </button>   
                     ):(
-                    <button className={`text-4xl duration-75 bg-green-600 active:bg-violet-700  shadow-md shadow-gray-900 text-white  rounded-full py-8 px-2 ${isLoading? 'animate-fadeIn':'opacity-0'} `} onClick={Hold}>
+                    <button className={`active:shadow-inner active:shadow-black  text-4xl duration-75 bg-green-600 active:bg-violet-700  shadow-md shadow-gray-900 text-white  rounded-full py-8 px-2 ${isLoading? 'animate-fadeIn':'opacity-0'} `} onClick={Hold}>
                         HOLD
                     </button>  
                     )}
